pipeline-registry/shopify-pipeline/v2025-07/514-labs/python/default/requirements.txt
python-dotenv>=1.0.1
requests>=2.31.0
httpx>=0.27.0
orjson>=3.9.15
structlog>=23.3.0
moose-cli==0.5.14
moose-lib==0.5.14
redis>=5.0.0
